requirements — копия.txt
Django~=3.2.6
pyzbar~=0.1.8
requests~=2.26.0
Pillow~=8.3.2